debug: move qemu to drivers
